renamed PrecedingLabelFunction to SucceedingLabelFunction
