+++ bark/world/evaluation/ltl/label_functions/succeeding_agent_label_function.cpp
@@ -6,21 +6,21 @@
 // This work is licensed under the terms of the MIT license.
 // For a copy, see <https://opensource.org/licenses/MIT>.
 
-#include "preceding_agent_label_function.hpp"
+#include "succeeding_agent_label_function.hpp"
 #include "bark/world/observed_world.hpp"
 
 using bark::world::AgentFrenetPair;
 using bark::world::AgentId;
 
-bark::world::evaluation::PrecedingAgentLabelFunction::
-    PrecedingAgentLabelFunction(const std::string& string,
+bark::world::evaluation::SucceedingAgentLabelFunction::
+    SucceedingAgentLabelFunction(const std::string& string,
                                 bool use_frac_param_from_world,
                                 double frac_lateral_offset)
     : MultiAgentLabelFunction(string),
       use_frac_param_from_world_(use_frac_param_from_world),
       frac_lateral_offset_(frac_lateral_offset) {}
 
-bool bark::world::evaluation::PrecedingAgentLabelFunction::EvaluateAgent(
+bool bark::world::evaluation::SucceedingAgentLabelFunction::EvaluateAgent(
     const bark::world::ObservedWorld& observed_world,
     const bark::world::objects::AgentPtr& other_agent) const {
   bool is_in_front = false;
