+++ bark/world/evaluation/ltl/label_functions/succeeding_agent_label_function.hpp
